ExROptionEditorを機能単位で分離し再レンダリングを最適化
- タブ選択UIを `ExRTabSelector` として独立
- カテゴリ一覧を `ExRCategoryList` として独立
- 個別のカテゴリを `CategoryAccordion` として独立させ、自身に関連する状態のみを監視するように修正
- カテゴリの開閉が他のアコーディオンやタブUIに影響を与えないように最適化
